requirements.txt
dvc==3.62.0
dvc-data==3.16.11
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
evidently==0.4.9
fastapi==0.116.1
fastapi-restful==0.6.0
mlflow==3.2.0
mlflow-skinny==3.2.0
mlflow-tracing==3.2.0
numpy==1.26.4
pandas==2.2.3
prometheus-client==0.19.0
psycopg2-binary==2.9.10
scikit-learn==1.6.1
uvicorn==0.35.0
python-dotenv==1.0.0
